requirements.txt
Flask==2.2.5
Flask-Login==0.6.3
Flask-SQLAlchemy==2.5.1
Werkzeug==2.2.3
numpy==1.24.3
scipy==1.10.1
opencv-python==4.7.0.72
Pillow==9.4.0
joblib==1.2.0
ultralytics==8.0.94